pyproject.toml
[project]
name = "mario"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "miniwdl>=1.12.1",
]
